Revert "to kotlin 1.3"
This reverts commit 5b821fc520c1f939089d5c78b2fe900ce19460b5.

+++ src/main/java/network/xyo/mod_bluetooth_kotlin/XyoBluetoothClient.kt
@@ -5,7 +5,8 @@ import android.bluetooth.BluetoothGatt
 import android.bluetooth.BluetoothManager
 import android.content.Context
 import android.os.ParcelUuid
-import kotlinx.coroutines.*
+import android.util.Log
+import kotlinx.coroutines.experimental.*
 import network.xyo.ble.devices.XYBluetoothDevice
 import network.xyo.ble.devices.XYCreator
 import network.xyo.ble.gatt.XYBluetoothError
@@ -19,8 +20,7 @@ import java.nio.ByteBuffer
 import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.collections.HashMap
-import kotlin.coroutines.resume
-import kotlin.coroutines.suspendCoroutine
+import kotlin.coroutines.experimental.suspendCoroutine
 
 
 /**
@@ -144,7 +144,7 @@ class XyoBluetoothClient(context: Context, device: BluetoothDevice?, hash : Int)
                                 if (waitForResponse) {
                                     logInfo("Going to read entire server response packet.")
 
-                                    delay(WAIT_FOR_RESPONSE_DELAY.toLong())
+                                    delay(WAIT_FOR_RESPONSE_DELAY)
                                     valueIn = readIncommoding()
                                 }
